Added support for pagination!
Pagination is now always disabled by default. If you add a "Page"
parameter into the request options you can request the specified page.
A -1 for the page queries the "last" page. If page is set to 0 it also
disables the pagination.

Beware of getting really long answers when not using pagination. We just
choose the disable pagination for default because in automatic
processing it is often desired to just not deal with paginations until
the result sets are huge or presented interactively.